backend/requirements.txt
# directly edited without requirments.in
apscheduler
asyncpg==0.30.0
beautifulsoup4
celery
chroma-hnswlib==0.7.6
chromadb==1.0.16
chromadb-ops==0.1.4
fastapi==0.115.6
fastapi-sso
ffmpeg-python
imutils
jwt==1.3.1
# konlpy==0.6.0
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-openai==0.3.29
langchain-tavily==0.2.12
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.3.45
litellm
llvmlite==0.44.0
nltk
numpy==1.26.4
openai==1.93.2
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
oracledb
pandas
pydantic[email]
python-dotenv
pycryptodome==3.21.0
python-multipart
python-pptx
pymupdf
pillow==11.1.0
pytz==2025.1
ragas==0.2.15
redis
requests
scipy==1.15.1
typing
uvicorn[standard]==0.34.0
